Split styles into editorialStyles, added Palette
